withAuthenticate.js works properly and feeds through App.js. Site views like normal.
