+++ insta-clone/src/components/authenticate/withAuthenticate.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+
+
+const withAuthenticate = RenderedComponent => 
+    class extends React.Component {
+        render() {
+            return <RenderedComponent />
+        }
+    }
+
+
+export default withAuthenticate;
